WIP: N-API migration - typed arrays, vsimem, utilities
61 -> 26 translation units still failing.

* utils/typed_array.cpp: rewritten on Napi::TypedArrayOf/Napi::Buffer;
  Nan::TypedArrayContents and Nan::NewBuffer are gone, the V8 constructors are
  reached through env.Global().
* gdal_memfile.cpp: the anonymous /vsimem/ buffer's weak callback is now
  napi_add_finalizer (the N-API equivalent), Buffer::HasInstance/Data/Length
  became Napi::Buffer, and the multi-line info.GetReturnValue().Set() call
  returns its value.
* utils/string_list.cpp: GetPropertyNames() and the string conversions.
* Nan::True()/Nan::False()/Nan::New<String>() rules; Local<String/Number/
  Boolean/Function> type names.

+++ src/gdal_dataset.cpp
@@ -855,7 +855,7 @@ GDAL_ASYNCABLE_GETTER_DEFINE(Dataset::geoTransformGetter) {
   job.rval = [](std::shared_ptr<double[]> transform, const GetFromPersistentFunc &) {
     if (transform == nullptr) return node_gdal::napi_env().Null().As<v8::Value>();
     Napi::Array result = Napi::Array::New(node_gdal::napi_env(), 6);
-    result.Set( 0, Napi::Number::New(node_gdal::napi_env(), transform.get()[0]));
+    result.Set(static_cast<uint32_t>(0), Napi::Number::New(node_gdal::napi_env(), transform.get()[0]));
     result.Set( 1, Napi::Number::New(node_gdal::napi_env(), transform.get()[1]));
     result.Set( 2, Napi::Number::New(node_gdal::napi_env(), transform.get()[2]));
     result.Set( 3, Napi::Number::New(node_gdal::napi_env(), transform.get()[3]));
